YAY-14 fix: rename Folder and Stories File for Filter Block
