+++ src/features/post-creator/index.ts
@@ -1,3 +1,3 @@
 export * from './constants/postCreatorConstants'
 export * from './hook/useModalStack'
-export * from './ui/modals/FilterModal/FiltersBlock'
+export * from './ui/modals/FilterBlock/FiltersBlock'
