+++ src/features/post-creator/ui/modals/FilterBlock/FiltersBlock.tsx
