+++ src/features/post-creator/ui/modals/FilterBlock/assets/Filter.jpg
